Update userCreation.ps1
Added try-catch check to check if a user already exists by the SamAccountName. Will need move from a splat object to a custom object for better use of variables.
Added console outputs for when a user already exists, and if not it'll create the user. Outputs when a user is created.
Added some foreground colors to console outputs to make things look better, and to make it more noticeable if a user already exists or is being created.
To fix: the try- check also outputs user information. This is unnecessary for the purpose of this script, but will be left in until a better solution is implemented.
Tested and working. Committing to main, will make modifications and push the next update as a pull to start.